Increased translation distance for notes when they unmount, so they fly completely off the screen.
Refactored state management for, keeping track of active notes, to use the note's ID instead of the whole note object.

Created a new selector for the NoteEditModal's textarea container to cap the height and introduce scrolling behaviour when the size of the modal is increased to its maximum height.  Styled the scroll bar for Firefox so that is looks better.

Decreased height of the Notes container, so it fits in the viewport instead of overflowing beneath.