Set up state for changing the color of the note being edited.  Had to do a major refactor of several components and created some buttons to manage saving and closing the modal.
Gave color dots a slight drop shadow, so they are more visible on the note header.

Created an enum for colors.

Refactored modal context.

Fixed several bugs along the way.

Created a new component, ActionButton, which takes some text to display, an action to perform on click, and a base color for customizing it.

Refactored state layout for the NoteEditModal.