Refactored Draggable component to use children instead of render props.
Created a new component called DraggableDnD, which is the draggable for the react-beautiful-dnd library.  It will be used when the mobile layout is applied.

Changed how margin is applied to individual notes.

Added a prop to the Note component so a ref can be passed to its div for the DraggableDnd to target.

Refactored note rendering in the Notes component.